requirements.txt
pillow
lxml
bs4
beautifulsoup4
matplotlib
tqdm
requests
requests-doh
requests-toolbelt
numpy
scipy
pandas
scikit-learn
torch
torchaudio
torchvision
transformers
tokenizers
manga-py
mangadex
mloader
boto3